Initial commit of flexbar_trim Galaxy tool
Wrapper for trimming TSO and Illumina adapters using 2 flexbar
commands connected by a pipe.

Note: Because flexbar 3.4.0 is not installed on the cluster yet, the
flexbar commands are not executed, but echo-ed. The inputs are copied
into the outputs.

This version wraps all the currently used flexbar options that were
optimized for trimming TSO and Illumina adapters.

TODO: Wrap the rest of the options.

TODO: Enable inherit_format* macros: waiting for
https://github.com/galaxyproject/galaxy/pull/6253 . As a temporary
workaround, using default format="fastqsanger" for the output files.